Rename and refactor package "appUser" -> "user".
