+++ src/main/java/com/example/demo/AppStartupRunner.java
@@ -1,7 +1,7 @@
 package com.example.demo;
 
-import com.example.demo.domain.appUser.User;
-import com.example.demo.domain.appUser.UserService;
+import com.example.demo.domain.user.User;
+import com.example.demo.domain.user.UserService;
 import com.example.demo.domain.authority.Authority;
 import com.example.demo.domain.authority.AuthorityRepository;
 import com.example.demo.domain.role.Role;
