+++ src/main/java/com/example/demo/domain/user/UserRepository.java
@@ -1,10 +1,7 @@
-package com.example.demo.domain.appUser;
+package com.example.demo.domain.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Modifying;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.UUID;
