+++ src/main/java/com/example/demo/domain/user/dto/PrivateUserDTO.java
@@ -1,11 +1,10 @@
-package com.example.demo.domain.appUser.dto;
+package com.example.demo.domain.user.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import javax.persistence.Column;
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
