+++ src/main/java/com/example/demo/domain/user/dto/PublicUserDTO.java
@@ -1,4 +1,4 @@
-package com.example.demo.domain.appUser.dto;
+package com.example.demo.domain.user.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
