+++ src/main/java/com/example/demo/domain/userProfile/UserProfileController.java
@@ -1,12 +1,10 @@
 package com.example.demo.domain.userProfile;
 
-import com.example.demo.domain.appUser.UserMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import javax.management.InstanceAlreadyExistsException;
 import javax.management.InstanceNotFoundException;
 import java.util.Collection;
 import java.util.Optional;
