+++ src/main/java/com/example/demo/domain/userProfile/UserProfileRepository.java
@@ -1,6 +1,5 @@
 package com.example.demo.domain.userProfile;
 
-import com.example.demo.domain.appUser.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
