+++ src/main/java/com/example/demo/domain/userProfile/UserProfileServiceImpl.java
@@ -1,7 +1,7 @@
 package com.example.demo.domain.userProfile;
 
-import com.example.demo.domain.appUser.UserRepository;
-import com.example.demo.domain.appUser.UserServiceImpl;
+import com.example.demo.domain.user.UserRepository;
+import com.example.demo.domain.user.UserServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
